Deleted all Console.WriteLine, fixed some bugs with program closing on some window close, multiple acc cannot be added if they are same

+++ EmailCheck/WindowsFormsLangs/AddAccount.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Net;
 using System.Net.Mail;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
@@ -22,34 +23,49 @@ namespace EmailCheck
         {
             InitializeComponent();
             this.chooseAccountWindow = chooseAccountWindow;
+            chooseAccountWindow.Enabled = false;
+            this.FormClosed += ClosedHandler;
             Button_AddAccount.Enabled = false;
             
         }
 
         private async void Button_AddAccount_Click(object sender, EventArgs e)
         {
+            RichTextBox_Error.Text = "";
             bool exceptionThrown = false;
-            try
+            bool emailExists = false;
+            foreach (User user in chooseAccountWindow.mainWindow.allUsers)
             {
-                List<MimeKit.MimeMessage> smtp = await CheckEmail.FetchAllMailbotMessages(this.TextBox_Email.Text, this.TextBox_Password.Text);
+                if (Regex.Replace(user.GetEmail(), @"\s+", "").Equals(Regex.Replace(this.TextBox_Email.Text, @"\s+", ""))) {
+                    emailExists = true;
+                }
             }
-            catch (Exception excep)
+            if (!emailExists)
             {
-                Warning warning = new Warning("Ivedėte neteisingą el. paštą arba slaptažodį");
+                    bool x = await CheckEmail.Login(this,this.TextBox_Email.Text, this.TextBox_Password.Text);
+                if (x)
+                {
+                    Warning warning = new Warning(this, "Ivedėte neteisingą el. paštą arba slaptažodį");
+                    warning.Show();
+                    RichTextBox_Error.Text = "Ivedėte neteisingą el. paštą arba slaptažodį";
+                    exceptionThrown = true;
+                }
+                if (!exceptionThrown)
+                {
+                    String email = StringCipher.Encrypt(this.TextBox_Email.Text, "125847elpu68795");
+                    String password = StringCipher.Encrypt(this.TextBox_Password.Text, "125847elpu62195");
+                    string fileDirectory = Directory.GetParent(this.workingDirectory).Parent.FullName + @"\ConfidentialInformation\UsersEmailCredentials.txt";
+                    File.WriteAllText(fileDirectory, File.ReadAllText(fileDirectory) + email + " \n " + password + " ");
+                    chooseAccountWindow.mainWindow.allUsers.Add(new User(this.TextBox_Email.Text, this.TextBox_Password.Text));
+                    chooseAccountWindow.ListBox_Accounts.Items.Add(this.TextBox_Email.Text);
+                    chooseAccountWindow.Enabled = true;
+                    this.Dispose();
+                }
+            }
+            else
+            {
+                Warning warning = new Warning(this,"Įrašytas el.paštas jau egzistuoja");
                 warning.Show();
-                RichTextBox_Error.Text = "Ivedėte neteisingą el. paštą arba slaptažodį";
-                exceptionThrown = true;
-            }
-            Console.WriteLine(exceptionThrown);
-            if (!exceptionThrown)
-            {
-                String email = StringCipher.Encrypt(this.TextBox_Email.Text, "125847elpu68795");
-                String password = StringCipher.Encrypt(this.TextBox_Password.Text, "125847elpu62195");
-                string fileDirectory = Directory.GetParent(this.workingDirectory).Parent.FullName + @"\ConfidentialInformation\UsersEmailCredentials.txt";
-                File.WriteAllText(fileDirectory, File.ReadAllText(fileDirectory) + email + " \n " + password + " ");
-                chooseAccountWindow.mainWindow.allUsers.Add(new User(this.TextBox_Email.Text, this.TextBox_Password.Text));
-                chooseAccountWindow.ListBox_Accounts.Items.Add(this.TextBox_Email.Text);
-                this.Dispose();
             }
         }
 
@@ -62,6 +78,10 @@ namespace EmailCheck
         {
             Button_AddAccount.Enabled = TextBox_Email.Text.Length == 0 || TextBox_Password.Text.Length == 0 ? false : true;
         }
- 
+        protected void ClosedHandler(object sender, EventArgs e)
+        {
+            chooseAccountWindow.Enabled = true;
+        }
+
     }
 }
